almost there

+++ db/migrate/20151102170423_create_positions.rb
@@ -3,7 +3,7 @@ class CreatePositions < ActiveRecord::Migration
     create_table :positions do |t|
       t.string :name
       t.date :expires_at
-      t.decimal :salary
+      t.float :salary
       t.text :contacts
 
       t.timestamps null: false
